Add CountdownTimer, HopeQuote, and SectionDivider components
- Introduced CountdownTimer component for displaying time left until a target date.
- Added HopeQuote component for displaying quotes with optional author attribution.
- Implemented SectionDivider component for visual separation with customizable styles.
- Updated page.tsx to utilize the new components for enhanced functionality and layout.
