+++ app/page.tsx
@@ -3,6 +3,9 @@
 import { Button } from "@/components/ui/button"
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@/components/ui/card"
 import { Badge } from "@/components/ui/badge"
+import { CountdownTimer } from "@/components/countdown-timer"
+import { HopeQuote } from "@/components/hope-quote"
+import { SectionDivider } from "@/components/section-divider"
 import {
   Clock,
   MapPin,
@@ -19,104 +22,6 @@ import {
   BookOpen,
 } from "lucide-react"
 import { useState, useEffect } from "react"
-
-function CountdownTimer({ targetDate }: { targetDate: Date }) {
-  const [timeLeft, setTimeLeft] = useState({
-    days: 0,
-    hours: 0,
-    minutes: 0,
-    seconds: 0,
-  })
-
-  useEffect(() => {
-    const timer = setInterval(() => {
-      const now = new Date().getTime()
-      const distance = targetDate.getTime() - now
-
-      if (distance > 0) {
-        setTimeLeft({
-          days: Math.floor(distance / (1000 * 60 * 60 * 24)),
-          hours: Math.floor((distance % (1000 * 60 * 60 * 24)) / (1000 * 60 * 60)),
-          minutes: Math.floor((distance % (1000 * 60 * 60)) / (1000 * 60)),
-          seconds: Math.floor((distance % (1000 * 60)) / 1000),
-        })
-      }
-    }, 1000)
-
-    return () => clearInterval(timer)
-  }, [targetDate])
-
-  return (
-    <div className="bg-gradient-to-r from-orange-400 to-yellow-400 text-white p-6 md:p-12 rounded-2xl md:rounded-3xl shadow-2xl mb-8 transform hover:scale-105 transition-transform duration-300">
-      <h3 className="text-3xl md:text-5xl font-bold mb-6 md:mb-8 text-center animate-pulse">¡Faltan solo!</h3>
-      <div className="grid grid-cols-4 gap-2 md:gap-6 text-center">
-        <div className="md:bg-white/30 md:rounded-xl p-2 md:p-4 md:m-2 md:backdrop-blur-md md:border md:border-white/40 md:shadow-lg md:min-w-[100px]">
-          <div className="text-4xl md:text-5xl font-bold text-center leading-tight">{timeLeft.days}</div>
-          <div className="text-lg md:text-xl text-center mt-1 md:mt-2">Días</div>
-        </div>
-        <div className="md:bg-white/30 md:rounded-xl p-2 md:p-4 md:m-2 md:backdrop-blur-md md:border md:border-white/40 md:shadow-lg md:min-w-[100px]">
-          <div className="text-4xl md:text-5xl font-bold text-center leading-tight">{timeLeft.hours}</div>
-          <div className="text-lg md:text-xl text-center mt-1 md:mt-2">Horas</div>
-        </div>
-        <div className="md:bg-white/30 md:rounded-xl p-2 md:p-4 md:m-2 md:backdrop-blur-md md:border md:border-white/40 md:shadow-lg md:min-w-[100px]">
-          <div className="text-4xl md:text-5xl font-bold text-center leading-tight">{timeLeft.minutes}</div>
-          <div className="text-lg md:text-xl text-center mt-1 md:mt-2">Min</div>
-        </div>
-        <div className="md:bg-white/30 md:rounded-xl p-2 md:p-4 md:m-2 md:backdrop-blur-md md:border md:border-white/40 md:shadow-lg md:min-w-[100px]">
-          <div className="text-4xl md:text-5xl font-bold text-center leading-tight">{timeLeft.seconds}</div>
-          <div className="text-lg md:text-xl text-center mt-1 md:mt-2">Seg</div>
-        </div>
-      </div>
-    </div>
-  )
-}
-
-function HopeQuote({ quote, author, className = "" }: { quote: string; author?: string; className?: string }) {
-  return (
-    <div className={`relative ${className}`}>
-      <div
-        className="bg-yellow-50 p-6 shadow-lg transform rotate-1 border-l-4 border-yellow-400"
-        style={{
-          clipPath: "polygon(0% 0%, 95% 0%, 100% 15%, 98% 85%, 100% 100%, 5% 100%, 0% 85%, 2% 15%)",
-        }}
-      >
-        <p className="text-gray-800 font-medium italic text-lg leading-relaxed">"{quote}"</p>
-        {author && <p className="text-gray-600 text-sm mt-2 text-right">- {author}</p>}
-      </div>
-    </div>
-  )
-}
-
-function SectionDivider({ variant = "wave" }: { variant?: "wave" | "diagonal" }) {
-  if (variant === "diagonal") {
-    return (
-      <div className="relative h-16 overflow-hidden">
-        <div className="absolute inset-0 transform -skew-y-1" style={{ backgroundColor: "#fd9903" }} />
-      </div>
-    )
-  }
-
-  return (
-    <div className="relative h-16 overflow-hidden">
-      <svg className="absolute bottom-0 w-full h-16" viewBox="0 0 1200 120" preserveAspectRatio="none">
-        <path
-          d="M0,0V46.29c47.79,22.2,103.59,32.17,158,28,70.36-5.37,136.33-33.31,206.8-37.5C438.64,32.43,512.34,53.67,583,72.05c69.27,18,138.3,24.88,209.4,13.08,36.15-6,69.85-17.84,104.45-29.34C989.49,25,1113-14.29,1200,52.47V0Z"
-          opacity=".25"
-          fill="#fd9903"
-        />
-        <path
-          d="M0,0V15.81C13,36.92,27.64,56.86,47.69,72.05,99.41,111.27,165,111,224.58,91.58c31.15-10.15,60.09-26.07,89.67-39.8,40.92-19,84.73-46,130.83-49.67,36.26-2.85,70.9,9.42,98.6,31.56,31.77,25.39,62.32,62,103.63,73,40.44,10.79,81.35-6.69,119.13-24.28s75.16-39,116.92-43.05c59.73-5.85,113.28,22.88,168.9,38.84,30.2,8.66,59,6.17,87.09-7.5,22.43-10.89,48-26.93,60.65-49.24V0Z"
-          opacity=".5"
-          fill="#fd9903"
-        />
-        <path
-          d="M0,0V5.63C149.93,59,314.09,71.32,475.83,42.57c43-7.64,84.23-20.12,127.61-26.46,59-8.63,112.48,12.24,165.56,35.4C827.93,77.22,886,95.24,951.2,90c86.53-7,172.46-45.71,248.8-84.81V0Z"
-          fill="#fd9903"
-        />
-      </svg>
-    </div>
-  )
-}
 
 export default function FestivalEsperanza() {
   const festivalDate = new Date("2025-11-22T08:00:00")
